+++ src/app.py
@@ -30,16 +30,6 @@ app.config['UPLOAD_FOLDER'] = os.getenv('UPLOAD_FOLDER')
 MIGRATE = Migrate(app, db)
 db.init_app(app)
 
-# app = Flask(__name__)
-# #app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///" + os.path.join(BASEDIR, "test.db")
-# app.config["DEBUG"] = True
-# app.config["ENV"] = "development"
-# app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
-# app.config["SECRET_KEY"] = "secret-key"
-# app.config['JWT_SECRET_KEY'] = 'encrypt'
-# app.config['UPLOAD_FOLDER'] = os.getenv('UPLOAD_FOLDER')
-
-
 # Allow CORS requests to this API
 CORS(app)
