+++ src/front/js/component/Portafolio.jsx
@@ -2,8 +2,9 @@ import React from "react";
 import { useState } from "react";
 import { serviceDiseno, serviceMarketing, serviceIt } from "../../img/image.js";
 import { Carousel } from "react-bootstrap";
+import PropTypes from "prop-types";
 
-const Portafolio = () => {
+const Portafolio = props => {
 	const [index, setIndex] = useState(0);
 	const handleSelect = (selectedIndex, e) => {
 		setIndex(selectedIndex);
@@ -11,7 +12,7 @@ const Portafolio = () => {
 	return (
 		<Carousel activeIndex={index} onSelect={handleSelect} className="mb-3">
 			<Carousel.Item>
-				<img className="d-block w-100" src={serviceDiseno} alt="First slide" />
+				<img className="d-block w-100" src={props.img} alt="First slide" />
 				<Carousel.Caption>
 					<p />
 				</Carousel.Caption>
@@ -32,3 +33,7 @@ const Portafolio = () => {
 	);
 };
 export default Portafolio;
+
+Portafolio.propTypes = {
+	img: PropTypes.string
+};
